+++ CHANGELOG.md
@@ -9,6 +9,10 @@ Format based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/)
 * `--sources` to show definitions' source dictionary
 * `--from` to only show results from a given source dictionary
 
+### Fixed
+
+* Some flags not having help text
+
 ### Changed
 
 * Updated dependencies